Move db interactions to NoteRepository
